+++ src/Common/Configuration.php
@@ -5,13 +5,18 @@ use Robo\Config;
 
 trait Configuration
 {
+    private static function getClassKey($key)
+    {
+        return sprintf('%s.%s', get_called_class(), $key);
+    }
+
     public static function configure($key, $value)
     {
-        Config::set(__CLASS__.".$key", $value);
+        Config::set(static::getClassKey($key), $value);
     }
 
     protected function getConfigValue($key, $default = null)
     {
-        return Config::get(__CLASS__.".$key", $default);
+        return Config::get(static::getClassKey($key), $default);
     }
 }